Re-use the HTTP connection for requests to the RabbitMQ API
The purpose of re-using the connection is to avoid leaving behind many
connections in TIME_WAIT state (one per queue).

To be able to do that, it was necessary to replace uses of the urllib2 library
with the requests library. The former shipped with python, the latter needs to
be separately installed. Therefore this plugin gained a dependency, the python
requests library.